Use the min/max arguments as boundaries for the histograms
We used the --min/--max arguments as filters, but not as boundaries for
the final histograms (they were using the actual min/max values of the
events in the trace that matched the filter). This resulted in hard to
compare histograms from differents experiments. Now we use these values
for the output as well as the filtering.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>