Flag to override the tracer version
The tracer version check causes problems on some traces, for example
when a trace directory contains a kernel and a UST trace and we don't
want to make the analysis process the UST trace, or with the traces
generated by CTFWriter.
This flags on the CLI allows the user to skip the auto-check and
set the version manually.

This could also be useful to test the behaviour of the analyzes when a
field changes between tracer versions.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>